@@ -26,18 +26,18 @@
         "composer/ca-bundle": "^1.0",
         "composer/class-map-generator": "^1.0",
         "composer/metadata-minifier": "^1.0",
-        "composer/semver": "^3.0",
+        "composer/semver": "^3.2.5",
         "composer/spdx-licenses": "^1.5.7",
         "composer/xdebug-handler": "^2.0.2 || ^3.0.3",
         "justinrainbow/json-schema": "^5.2.11",
         "psr/log": "^1.0 || ^2.0 || ^3.0",
         "seld/jsonlint": "^1.4",
         "seld/phar-utils": "^1.2",
-        "symfony/console": "^5.4.11 || ^6.0.11",
-        "symfony/filesystem": "^5.4 || ^6.0",
-        "symfony/finder": "^5.4 || ^6.0",
-        "symfony/process": "^5.4 || ^6.0",
-        "react/promise": "^2.8",
+        "symfony/console": "^5.4.11 || ^6.0.11 || ^7",
+        "symfony/filesystem": "^5.4 || ^6.0 || ^7",
+        "symfony/finder": "^5.4 || ^6.0 || ^7",
+        "symfony/process": "^5.4 || ^6.0 || ^7",
+        "react/promise": "^2.8 || ^3",
         "composer/pcre": "^2.1 || ^3.1",
         "symfony/polyfill-php73": "^1.24",
         "symfony/polyfill-php80": "^1.24",
@@ -45,7 +45,7 @@
         "seld/signal-handler": "^2.0"
     },
     "require-dev": {
-        "symfony/phpunit-bridge": "^6.0",
+        "symfony/phpunit-bridge": "^6.4.1 || ^7.0.1",
         "phpstan/phpstan": "^1.9.3",
         "phpstan/phpstan-phpunit": "^1.0",
         "phpstan/phpstan-deprecation-rules": "^1",
@@ -65,7 +65,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-main": "2.5-dev"
+            "dev-main": "2.7-dev"
         },
         "phpstan": {
             "includes": [
@@ -75,12 +75,12 @@
     },
     "autoload": {
         "psr-4": {
-            "Composer\\": "src/Composer"
+            "Composer\\": "src/Composer/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Composer\\Test\\": "tests/Composer/Test"
+            "Composer\\Test\\": "tests/Composer/Test/"
         }
     },
     "bin": [
@@ -98,6 +98,7 @@
     },
     "support": {
         "issues": "https://github.com/composer/composer/issues",
-        "irc": "ircs://irc.libera.chat:6697/composer"
+        "irc": "ircs://irc.libera.chat:6697/composer",
+        "security": "https://github.com/composer/composer/security/policy"
     }
 }
