@@ -13,11 +13,11 @@
         "phpstan": "vendor/bin/phpstan analyse -l 6 src"
     },
     "require": {
-        "php": "^7.2 || ^8.0",
+        "php": "^7.4 || ^8.0",
         "ext-fileinfo": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5.8 || ^9.3",
+        "phpunit/phpunit": "^8.5.8 || ^9.3 || ^10.0",
         "phpstan/phpstan": "^0.12.68",
         "friendsofphp/php-cs-fixer": "^3.2"
     },
@@ -28,7 +28,7 @@
     },
     "config": {
         "platform": {
-            "php": "7.2.0"
+            "php": "7.4.0"
         }
     }
 }
