@@ -23,7 +23,10 @@
     "require-dev": {
         "symfony/http-kernel": "^5.4|^6.0",
         "symfony/serializer": "^5.4|^6.0",
-        "symfony/deprecation-contracts": "^2.1|^3"
+        "symfony/deprecation-contracts": "^2.5|^3"
+    },
+    "conflict": {
+        "symfony/deprecation-contracts": "<2.5"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\ErrorHandler\\": "" },
