@@ -15,13 +15,14 @@
 		}
 	],
 	"require": {
-		"php": ">=7.2 <8.3",
+		"php": ">=8.0 <8.4",
 		"ext-session": "*",
 		"ext-json": "*"
 	},
 	"require-dev": {
 		"nette/utils": "^3.0",
 		"nette/di": "^3.0",
+		"nette/http": "^3.0",
 		"nette/mail": "^3.0",
 		"nette/tester": "^2.2",
 		"latte/latte": "^2.5",
@@ -42,7 +43,7 @@
 	},
 	"extra": {
 		"branch-alias": {
-			"dev-master": "2.9-dev"
+			"dev-master": "2.10-dev"
 		}
 	}
 }
