@@ -16,18 +16,18 @@
         }
     ],
     "require": {
-        "php": "^7.1.3 || ^8.0",
+        "php": "^7.2.5 || ^8.0",
         "ext-pcre": "*",
-        "graham-campbell/result-type": "^1.0.2",
-        "phpoption/phpoption": "^1.8",
-        "symfony/polyfill-ctype": "^1.23",
-        "symfony/polyfill-mbstring": "^1.23.1",
-        "symfony/polyfill-php80": "^1.23.1"
+        "graham-campbell/result-type": "^1.1.2",
+        "phpoption/phpoption": "^1.9.2",
+        "symfony/polyfill-ctype": "^1.24",
+        "symfony/polyfill-mbstring": "^1.24",
+        "symfony/polyfill-php80": "^1.24"
     },
     "require-dev": {
         "ext-filter": "*",
-        "bamarni/composer-bin-plugin": "^1.4.1",
-        "phpunit/phpunit": "^7.5.20 || ^8.5.30 || ^9.5.25"
+        "bamarni/composer-bin-plugin": "^1.8.2",
+        "phpunit/phpunit":"^8.5.34 || ^9.6.13 || ^10.4.2"
     },
     "autoload": {
         "psr-4": {
@@ -54,7 +54,7 @@
             "forward-command": true
         },
         "branch-alias": {
-            "dev-master": "5.5-dev"
+            "dev-master": "5.6-dev"
         }
     }
 }
