@@ -13,6 +13,12 @@ if(!empty($links)) {
     }
 }
 if(empty($body)) return;
+
+$params = $modx->event->params;
+$col_lg = isset($params["col_lg"]) ? (int)$params["col_lg"] : 6;
+$menu_index = isset($params["col_lg"]) ? (int)$params["menu_index"] : 10;
+$hide_menu = isset($params["hide_menu"]) ? (int)$params["hide_menu"] : 0;
+
 $body .= "<style>
 .widgets #about .wm_buttons {
     display: -ms-flexbox;
@@ -25,6 +31,7 @@ $body .= "<style>
 
 .widgets #about .wm_button {
     max-width: 50%;
+    min-width: 120px;
     flex: 0 0 50%
 }
 
@@ -83,13 +90,13 @@ $body .= "<style>
 }</style>";
 $body = '<div class="wm_buttons card-body">' . $body . '</div>';
 $widgets['about'] = [
-    'menuindex' => '11',
+    'menuindex' => $menu_index,
     'id'        => 'about',
-    'cols'      => 'col-lg-6',
+    'cols'      => 'col-lg-' . $col_lg,
     'icon'      => 'fa-info-circle',
     'title'     => '[%about_title%]',
     'body'      => $body,
-    'hide'      => 0
+    'hide'      => $hide_menu
 ];
 
 $modx->event->addOutput(serialize($widgets));
