@@ -14,13 +14,13 @@
 
 $_lang['pluginname'] = 'Updater';
 $_lang['system_update'] = 'Systeem update';
-$_lang["cms_outdated_msg"] = 'Het Content Management System is verouderd. Om deze bij te werken neem contact op met de website ontwikkelaar. Huidige versie';
-$_lang['bkp_before_msg'] = 'We raden het maken van een back-up aan voor de upgrade van het systeem. De update wordt uitgevoerd op eigen risico!';
-$_lang['updateButton_txt'] = 'Update naar nieuwe version_compare';
+$_lang["cms_outdated_msg"] = 'Er is een nieuwe versie van het systeem beschikbaar:';
+$_lang['bkp_before_msg'] = 'Raadpleeg eerst uw websitebeheerder voor het uitvoeren van een update!';
+$_lang['updateButton_txt'] = 'Updaten naar versie';
 //Error Messages
 $_lang['error_curl'] = 'Het is noodzakelijk om de CURL in PHP in te schakelen';
 $_lang['error_zip'] = 'Het is noodzakelijk om de ZIP in PHP in te schakelen';
 $_lang['error_openssl'] = 'De OpenSSL functie in PHP moet worden ingeschakelend';
 $_lang['error_overwrite'] = 'Bestanden kunnen niet worden overschreven';
 
-$_lang['artisan_update'] = 'Om bij te werken, moet u de console-opdracht vanuit de map uitvoeren <b>core</b>: <b>php artisan make:site update</b>';
+$_lang['artisan_update'] = 'Voor het updaten van het systeem moet u de console-opdracht vanuit de map <b>/core</b> uitvoeren: <b>php artisan make:site update</b>';
