@@ -48,7 +48,10 @@ return [
             'driver' => 'file',
             'path' => EVO_STORAGE_PATH . 'cache/',
         ],
-
+        'rss' => [
+            'driver' => 'file',
+            'path' => EVO_STORAGE_PATH . 'rss/',
+        ],
         'memcached' => [
             'driver' => 'memcached',
             'persistent_id' => env('MEMCACHED_PERSISTENT_ID'),
