@@ -889,7 +889,7 @@ $_lang["search_item_deleted"] = 'Темата е изтрита';
 $_lang["search_results"] = 'Резултати от търсенето';
 $_lang["search_results_returned_desc"] = 'Описание';
 $_lang["search_results_returned_id"] = 'ID';
-$_lang["search_results_returned_msg"] = 'Търсенето откри <b>%s</b> документа. В случай, че върнатите резултати са много, опитайте се да детайлизирате търсенето. Двете най-леви колони ви дават възможност да намерите документа в дървото или да го разгледате. Двете най-десни колони, съответно, показват дали документа е бил изтрит и неговия статус - публикуван или непубликуван.<p />';
+$_lang["search_results_returned_msg"] = 'Търсенето откри <b>%s</b> документа. В случай, че върнатите резултати са много, опитайте се да детайлизирате търсенето.';
 $_lang["search_results_returned_title"] = 'Заглавие';
 $_lang["search_view_docdata"] = 'Преглед на тема';
 $_lang["security"] = 'Сигурност';
