@@ -889,7 +889,7 @@ $_lang["search_item_deleted"] = 'Dette er blevet slettet';
 $_lang["search_results"] = 'Søgeresultater';
 $_lang["search_results_returned_desc"] = 'Beskrivelse';
 $_lang["search_results_returned_id"] = 'ID';
-$_lang["search_results_returned_msg"] = 'Søgningen resulterede i <b>%s</b> ressourcer. Prøv at søge mere specifikt hvis der er mange resultater. I de to kolonner til venstre kan man se eller finde den enkelte ressouce i website træet. I de to kolonner til højre kan man se om ressourcen er slettet og dennes status<p />';
+$_lang["search_results_returned_msg"] = 'Søgningen resulterede i <b>%s</b> ressourcer. Prøv at søge mere specifikt hvis der er mange resultater.';
 $_lang["search_results_returned_title"] = 'Titel';
 $_lang["search_view_docdata"] = 'Se denne';
 $_lang["security"] = 'Sikkerhed';
