@@ -889,7 +889,7 @@ $_lang["search_item_deleted"] = 'Denne posten har blitt fjernet';
 $_lang["search_results"] = 'Søkeresultat';
 $_lang["search_results_returned_desc"] = 'Beskrivelse';
 $_lang["search_results_returned_id"] = 'ID';
-$_lang["search_results_returned_msg"] = 'Ditt søk returnerte <b>%s</b> dokumenter. Prøv å skrive et mer spesifikt søk hvis du får mange treff. De to kolonnene lengst til venstre lar deg finne dokumentet i treet eller vise dokumentet. De to kolonnene til høyre viser om dokumentet har blitt fjernet og hvilken publiseringsstatus dokumentet har.<p />';
+$_lang["search_results_returned_msg"] = 'Ditt søk returnerte <b>%s</b> dokumenter. Prøv å skrive et mer spesifikt søk hvis du får mange treff.';
 $_lang["search_results_returned_title"] = 'Tittel';
 $_lang["search_view_docdata"] = 'Vis denne posten';
 $_lang["security"] = 'Sikkerhet';
