@@ -889,7 +889,7 @@ $_lang["search_item_deleted"] = 'Ta pozycja została usunięta';
 $_lang["search_results"] = 'Wyniki wyszukiwania';
 $_lang["search_results_returned_desc"] = 'Opis';
 $_lang["search_results_returned_id"] = 'ID';
-$_lang["search_results_returned_msg"] = 'Twoje kryteria wyszukiwania zwróciły <b>%s</b> dokumentów. Jeżeli wyników wyszukiwania jest bardzo dużo, spróbuj wpisać bardziej szczegółowe kryteria wyszukiwania. Dwie kolumny z lewej pozwolą ci znaleźć dokument w drzewie lub wyświetlić jego zawartość. Dwie kolumny po prawej pokażą odpowiednio, czy dokument został usunięty oraz czy jest opublikowany.<p />';
+$_lang["search_results_returned_msg"] = 'Twoje kryteria wyszukiwania zwróciły <b>%s</b> dokumentów. Jeżeli wyników wyszukiwania jest bardzo dużo, spróbuj wpisać bardziej szczegółowe kryteria wyszukiwania.';
 $_lang["search_results_returned_title"] = 'Tytuł';
 $_lang["search_view_docdata"] = 'Zobacz tą pozycję';
 $_lang["security"] = 'Użytkownicy ';
