@@ -889,7 +889,7 @@ $_lang["search_item_deleted"] = 'Este item foi apagado';
 $_lang["search_results"] = 'Resultados da procura';
 $_lang["search_results_returned_desc"] = 'Descrição';
 $_lang["search_results_returned_id"] = 'ID';
-$_lang["search_results_returned_msg"] = 'Os critérios da sua procura retornaram <b>%s</b> documentos. Se estiver a obter demasiados resultados, tente indicar uma procura mais específica. As duas colunas mais à esquerda permitem que encontre o documento na árvore ou ver o documento. As duas colunas mais à direita mostram, respectivamente, se um documento foi apagado e qual o seu estado de publicação<br />';
+$_lang["search_results_returned_msg"] = 'Os critérios da sua procura retornaram <b>%s</b> documentos. Se estiver a obter demasiados resultados, tente indicar uma procura mais específica.';
 $_lang["search_results_returned_title"] = 'Título';
 $_lang["search_view_docdata"] = 'Ver este item';
 $_lang["security"] = 'Segurança';
