@@ -889,7 +889,7 @@ $_lang["search_item_deleted"] = 'Denna post har tagits bort';
 $_lang["search_results"] = 'Sökresultat';
 $_lang["search_results_returned_desc"] = 'Beskrivning';
 $_lang["search_results_returned_id"] = 'ID';
-$_lang["search_results_returned_msg"] = 'Din sökning returnerade <b>%s</b> resurser. Försök att göra en mer specifik sökning om du får många träffar. De två kolumnerna längst till vänster låter dig hitta resursen i webbplatsträdet eller visa resursen. De två kolumnerna till höger visar om resursen har tagits bort och vilken publiceringsstatus den har.<p />';
+$_lang["search_results_returned_msg"] = 'Din sökning returnerade <b>%s</b> resurser. Försök att göra en mer specifik sökning om du får många träffar.';
 $_lang["search_results_returned_title"] = 'Titel';
 $_lang["search_view_docdata"] = 'Visa denna post';
 $_lang["security"] = 'Säkerhet';
