@@ -889,7 +889,7 @@ $_lang["search_item_deleted"] = '本项以被删除';
 $_lang["search_results"] = '搜索结果';
 $_lang["search_results_returned_desc"] = '描述';
 $_lang["search_results_returned_id"] = 'ID';
-$_lang["search_results_returned_msg"] = '搜索结果返回 <b>%s</b> 文件，如果返回结果太多，请限制您的搜索条件， The two left-most columns will allow you to find the document in the tree or view the document. The two right most columns show, respectively, if a document has been deleted and what it\'s published status is.<p />';
+$_lang["search_results_returned_msg"] = '搜索结果返回 <b>%s</b> 文件，如果返回结果太多，请限制您的搜索条件，';
 $_lang["search_results_returned_title"] = '标题';
 $_lang["search_view_docdata"] = '查阅本项';
 $_lang["security"] = '保密机密';
