@@ -55,7 +55,7 @@ class RefreshSite extends AbstractController implements ManagerTheme\PageControl
 
         $count = $query->count();
 
-        $query->update(['published' => 1]);
+        $query->update(['published' => 1, 'pub_date' => 0]);
 
         return $count;
     }
@@ -66,7 +66,7 @@ class RefreshSite extends AbstractController implements ManagerTheme\PageControl
 
         $count = $query->count();
 
-        $query->update(['published' => 0]);
+        $query->update(['published' => 0, 'unpub_date' => 0]);
 
         return $count;
     }
