@@ -24,7 +24,7 @@ class Router extends IlluminateRouter implements BindingRegistrar, RegistrarCont
             $this->addRoute(['GET', 'POST', 'PUT', 'DELETE', 'PATCH', 'OPTIONS'], '{any}', function () {
                 EvolutionCMS()->executeParser();
                 exit();
-            })->where('any', '.*')->fallback();
+            })->where('any', '.*')->middleware('web')->fallback();
 
             $this->hasFallbackToParser = true;
         }
