@@ -39,7 +39,7 @@ document.getElementById('elements-select').onchange = function(e) {
       });
     }
   };
-  xhr.send(request_key + '[ajax]=1&' + request_key + '[task]=categorize_load_elements&' + request_key + '[elements]=' + e.target.value);
+  xhr.send(request_key + '[ajax]=1&_token=' + token + '&' + request_key + '[task]=categorize_load_elements&' + request_key + '[elements]=' + e.target.value);
 };
 
 /**
