@@ -58,7 +58,7 @@
         var stay = document.getElementById('stay');
         // Trigger unlock
         if ((stay && stay.value !== '2') || !form_save) {
-          var url = '<?php echo MODX_MANAGER_URL; ?>?a=67&type={{ $elementType }}&id={{ $data->getKey() }}&o=' + Math.random();
+          var url = '<?php echo MODX_MANAGER_URL; ?>?a=67&type={{ $elementType }}&id={{ $data->getKey() }}&_token={{ csrf_token() }}&o=' + Math.random();
           if (navigator.sendBeacon) {
             navigator.sendBeacon(url)
           } else {
